Move drawer contents into component
