+++ src/components/OverviewDrawer/index.jsx
@@ -1,9 +1,8 @@
-import React, { useContext, useState } from "react";
-import { OverviewSearch } from "./OverviewSearch";
-import { Drawer, makeStyles, Hidden, Button } from "@material-ui/core";
+import React, { useContext } from "react";
+import { Drawer, makeStyles, Hidden } from "@material-ui/core";
 import { SpacedButton } from "../Material";
-import { OverviewList } from "./OverviewList";
 import { FirebaseAppContext } from "../../context/FirebaseContext";
+import { OverviewContent } from "./OverviewContent";
 
 const drawerWidth = 380;
 
@@ -55,17 +54,11 @@ const useStyles = makeStyles((theme) => ({
 const OverviewDrawer = (props) => {
   const classes = useStyles();
 
-  const { firebase, user } = useContext(FirebaseAppContext);
+  const { user } = useContext(FirebaseAppContext);
   const [state, setState] = React.useState({
     left: false,
   });
   const [isManager, setIsManager] = React.useState(false);
-  const [searchTerms, setSearchTerms] = React.useState([]);
-
-  const handleSearch = (val) => {
-    setSearchTerms(val);
-    // should pass the search upwards again, so I can pass searchTerms into component with React.clone in indexjsx of this module
-  };
 
   React.useEffect(() => {
     if (
@@ -90,8 +83,6 @@ const OverviewDrawer = (props) => {
   const renderDrawer = () => {
     // TODO: working list -> now mutates the clicked on person to whoever I am currently on
     // TODO: when saving on /create navigate to link/thatId
-    // TODO: move those components into a component to avoid repeating them
-    // TODO: persist component on mobile when toggling - searchterms and data reset and re-requested
     return (
       <>
         <Hidden lgUp>
@@ -113,25 +104,11 @@ const OverviewDrawer = (props) => {
                 paper: classes.drawerPaper,
               }}
             >
-              <div className={classes.drawerContent}>
-                <div className={classes.sticky}>
-                  <SpacedButton
-                    href="/creator"
-                    variant="contained"
-                    color="primary"
-                    marginBottom={2}
-                  >
-                    Add Resume
-                  </SpacedButton>
-                  <OverviewSearch handleSearch={handleSearch} />
-                </div>
-                <OverviewList
-                  firebase={firebase}
-                  user={user}
-                  searchTerms={searchTerms}
-                  query={firebase.firestore().collection("resumes")}
-                />
-              </div>
+              <OverviewContent
+                rootClass={classes.drawerContent}
+                actionClass={classes.sticky}
+                toggleDrawer={toggleDrawer}
+              />
             </Drawer>
           </Hidden>
           <Hidden lgUp>
@@ -148,34 +125,12 @@ const OverviewDrawer = (props) => {
                 keepMounted: true, // Better open performance on mobile.
               }}
             >
-              <div className={classes.drawerContentMobile}>
-                <div className={classes.sticky}>
-                  <SpacedButton
-                    href="/creator"
-                    variant="contained"
-                    color="primary"
-                    marginBottom={2}
-                  >
-                    Add Resume
-                  </SpacedButton>
-                  <SpacedButton
-                    onClick={toggleDrawer("left", false)}
-                    variant="contained"
-                    color="primary"
-                    marginBottom={2}
-                    marginLeft={2}
-                  >
-                    Close
-                  </SpacedButton>
-                  <OverviewSearch handleSearch={handleSearch} />
-                </div>
-                <OverviewList
-                  firebase={firebase}
-                  user={user}
-                  searchTerms={searchTerms}
-                  query={firebase.firestore().collection("resumes")}
-                />
-              </div>
+              <OverviewContent
+                rootClass={classes.drawerContentMobile}
+                actionClass={classes.sticky}
+                toggleDrawer={toggleDrawer}
+                isMobile={true}
+              />
             </Drawer>
           </Hidden>
           <div className={classes.content}>{props.children}</div>
